pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "patricia capitanelli"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph (>=0.6.6,<0.7.0)",
    "langchain-openai (>=0.3.32,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
     

    "pandas (>=2.0,<3.0)",
    "openpyxl (>=3.1,<4.0)",
    "langchain-community (>=0.3.10,<0.4.0)",
    "duckduckgo-search (>=6.3,<7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
